Migrate NFont to SDL3.
These changes allow for NFont to work with SDL3 and SDL3_ttf.

It requires SDL_FontCache to also be migrated, and luckily there is at
least one fork out there that has done so.

I am not sure what to do with the copyright and NFont version, but those
should probably be updated.